Add Swagger and Liquibase to Truck Category microservice

+++ truck-category-service/src/main/java/com/gotruck/truckcategoryservice/dto/TruckCategoryDTO.java
@@ -1,9 +1,12 @@
 package com.gotruck.truckcategoryservice.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.*;
 import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -15,23 +18,24 @@ public class TruckCategoryDTO {
 
     @NotNull(message = "Max load capacity must not be null")
     @Positive(message = "Max load capacity must be positive")
-    private double maxLoadCapacity;
+    private BigDecimal maxLoadCapacity;
 
     @NotNull(message = "Cargo area width must not be null")
     @Positive(message = "Cargo area width must be positive")
-    private double cargoAreaWidth;
+    private BigDecimal cargoAreaWidth;
 
     @NotNull(message = "Cargo area length must not be null")
     @Positive(message = "Cargo area length must be positive")
-    private double cargoAreaLength;
+    private BigDecimal cargoAreaLength;
 
     @NotNull(message = "Cargo area height must not be null")
     @Positive(message = "Cargo area height must be positive")
-    private double cargoAreaHeight;
+    private BigDecimal cargoAreaHeight;
 
     @Positive(message = "Cargo cubic volume must be positive")
-    private double cargoCubicVolume;
+    private BigDecimal cargoCubicVolume;
 
+//    @NotNull(message = "Truck name must not be null")
     private Long truckNameId;
 
 }
